requirements.txt
# -----------------------------
# Core ML / Diffusion stack
# -----------------------------
torch
diffusers>=0.20.0
transformers
accelerate
safetensors

# -----------------------------
# Data processing & utilities
# -----------------------------
numpy
Pillow
tqdm
huggingface_hub

# -----------------------------
# (Optional but commonly present)
# Uncomment if needed later
# -----------------------------
# scipy
# matplotlib
